Fixed Date Display and made one uniform helper function
