+++ client/src/pages/BusinessDivisions/BusinessDivisionsListAndEdit.tsx
@@ -23,6 +23,7 @@ import {
   getFieldChecks,
   checkForDuplicates,
 } from '../../utils/validationUtil'
+import { formatDate } from "../../utils/helperFunctionsUtil";
 
 const BusinessDivisionsListAndEdit: React.FC = () => {
     const [activeTab, setActiveTab] = useState('/planning-list')
@@ -352,14 +353,6 @@ const BusinessDivisionsListAndEdit: React.FC = () => {
     const handleNewRegistrationClick = () => {
       navigate('/business-divisions-registration');
     };
-
-    const formatDate = (dateString) => {
-      const date = new Date(dateString)
-      const month = String(date.getMonth() + 1).padStart(2, '0') // Get month (0-indexed, so +1)
-      const day = String(date.getDate()).padStart(2, '0') // Get day
-      const year = date.getFullYear() // Get full year
-      return `${month}/${day}/${year}`
-    }
 
   return (
     <div className='BusinessDivisionsListAndEdit_wrapper'>
